Move deleteInProgressForm to shared folder
