+++ src/applications/dependents/dependents-verification/util/index.js
@@ -2,22 +2,7 @@ import environment from 'platform/utilities/environment';
 import { apiRequest } from 'platform/utilities/api';
 import { transformForSubmit } from 'platform/forms-system/src/js/helpers';
 import recordEvent from 'platform/monitoring/record-event';
-import { removeFormApi } from 'platform/forms/save-in-progress/api';
 import { hideDependentsWarning } from '../../shared/utils';
-
-export async function deleteInProgressForm(formId) {
-  return removeFormApi(formId)
-    .then(() => {
-      recordEvent({
-        event: 'dependents-verification-delete-in-progress-form-success',
-      });
-    })
-    .catch(() => {
-      recordEvent({
-        event: 'dependents-verification-delete-in-progress-form-failure',
-      });
-    });
-}
 
 export function transform(formConfig, form) {
   const formData = transformForSubmit(formConfig, form);
